장바구니

+++ src/main/webapp/user/userList.jsp
@@ -24,13 +24,26 @@ text-align: center;
 			<h3>회원 목록</h3>
 			<div class="열1">
 			<c:forEach var="user" items="${userList }">
-			<div class="열2"><a href="userViewAction.u?id=${user.id }">${user.id }</a></div>
+			<div class="열2" ><a href="userViewAction.u?id=${user.id }">${user.id }
+			<c:if test="${user.user_bye }">
+			(탈퇴회원)
+			</c:if>
+			</a></div>
 			<div class="열2"><a href="userModifyForm.u?id=${user.id }">수정</a></div>
-			<div class="열2"><a href="userDeleteAction.u?id=${user.id }">삭제</a></div>
+			<div class="열2"><a href="userDeleteAction.u?id=${user.id }" onclick="confirmDelete('${user.id}' ,event)">삭제</a></div>
 			</c:forEach>
 			</div>
 		</div>
 	</div>
 </div>
+<script>
+function confirmDelete(userId, event) {
+	  event.preventDefault(); // 이벤트의 기본 동작을 취소
+
+	  if (confirm("정말 삭제하시겠습니까?")) {
+	    location.href = "userDeleteAction.u?id=" + userId;
+	  }
+	}
+</script>
 </body>
 </html>
